LPD-59495 Rename FDSPropsTransformer to props_transformer, so we follow the same pattern in every folder
