+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/categorization/tags/MergeTagsModal.tsx
@@ -17,7 +17,7 @@ import React, {useEffect, useState} from 'react';
 import SpaceSticker from '../../../common/components/SpaceSticker';
 import ApiHelper from '../../../common/services/ApiHelper';
 import {LogoColor} from '../../../common/types/Space';
-import {executeAsyncItemAction} from '../../FDSPropsTransformer/utils/executeAsyncItemAction';
+import {executeAsyncItemAction} from '../../props_transformer/utils/executeAsyncItemAction';
 
 type Tag = {
 	label: string;
