+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/modal/CreationModalContent.tsx
@@ -12,8 +12,8 @@ import React from 'react';
 import {FieldPicker, FieldText} from '../../common/components/forms';
 import {required, validate} from '../../common/components/forms/validations';
 import {AssetLibrary} from '../../common/types/AssetLibrary';
-import {AssetData} from '../FDSPropsTransformer/actions/createAssetAction';
-import {FolderData} from '../FDSPropsTransformer/actions/createFolderAction';
+import {AssetData} from '../props_transformer/actions/createAssetAction';
+import {FolderData} from '../props_transformer/actions/createFolderAction';
 
 type Props = {
 	action: AssetData['action'] | FolderData['action'];
